create auth module to login users
