+++ src/modules/auth/middleware/auth.middleware.ts
@@ -1,8 +1,7 @@
 import { Request, Response, NextFunction } from 'express';
-import { verifyToken } from '../../jwt.util';
+import { verifyToken } from '../application/utils/jwt.util';
 import { JwtPayload } from 'jsonwebtoken';
 
-// Extensión de la interfaz Request para que ts detecte el user en el request
 declare module 'express-serve-static-core' {
   interface Request {
     user?: string | JwtPayload;
@@ -10,8 +9,7 @@ declare module 'express-serve-static-core' {
 }
 
 const authMiddleware = (req: Request, res: Response, next: NextFunction) => {
-  const token = req.headers['authorization'].split(' ')[1];
-
+  const token = req.headers['cookie'].split('=')[1];
   if (!token) {
     return res.status(401).send('Access denied. No token provided.');
   }
