+++ src/modules/auth/routes/auth.router.ts
@@ -1,0 +1,8 @@
+import { Router } from 'express';
+import { AuthController } from '../controller/auth.controller';
+
+const authNavigation = Router();
+
+authNavigation.post('/login', AuthController.login);
+
+export default authNavigation;
