+++ src/modules/user/application/service/user.service.ts
@@ -3,8 +3,6 @@ import { UserRepository } from '../../infrastructure/persistence/user.repository
 import bcrypt from 'bcryptjs';
 import { CreateUserRequestDto } from '../dto/request/create-user.dto';
 import { newUserMapper } from '../utils/user.utils';
-import { generateToken } from '@/modules/jwt.util';
-import { LoginResponseDto } from '../dto/response/login.response.dto';
 
 export const UserService = {
   async getUserList() {
@@ -32,33 +30,7 @@ export const UserService = {
   async deleteUser(id: string) {
     return await UserRepository.deleteUser(id);
   },
-  async authenticateUser(
-    email: string,
-    password: string,
-  ): Promise<LoginResponseDto> {
-    const user = await UserRepository.getUserByEmail(email);
-    if (!user) {
-      throw new Error('User not found');
-    }
-
-    const isPasswordValid = await bcrypt.compare(password, user.password);
-
-    if (!isPasswordValid) {
-      throw new Error('Invalid password');
-    }
-
-    const token = generateToken({
-      email: user.email,
-      name: user.name,
-    });
-
-    const response = new LoginResponseDto();
-    response.accessToken = token;
-    response.expiresIn = '1h';
-    response.user = {
-      email: user.email,
-      name: user.name,
-    };
-    return response;
+  async getUserByEmail(email: string) {
+    return await UserRepository.getUserByEmail(email);
   },
 };
